requirements.txt
# Python requirements for DoomTeach
openai
python-dotenv
requests
Pillow
rembg
duckduckgo-search
pydub
fakeyou
moviepy==1.0.3
asyncio
aiolimiter
faster-whisper
google-adk
